Move JsonPrinter into new package.
