+++ test/shared/src/main/scala/zio/test/results/ExecutionEventJsonPrinter.scala
@@ -1,99 +1,90 @@
-package zio.test
+package zio.test.results
 
-import zio.interop.javaz
-import zio.stacktracer.TracingImplicits.disableAutoTrace
-import zio._
+import zio.test.{
+  ExecutionEvent,
+  ExecutionEventPrinter,
+  TestAnnotationMap,
+  TestAnnotationRenderer,
+  TestFailure,
+  TestSuccess
+}
+import zio.{Scope, Trace, ZIO, ZLayer, ZOutputStream}
 
 import java.io
 import java.io.IOException
-import java.net.{URI, URL}
-import java.nio.channels.CompletionHandler
-import java.nio.charset.StandardCharsets
 import java.nio.file.Path
-import java.util.concurrent.{CompletableFuture, CompletionStage, Future}
 
 object ExecutionEventJsonPrinter {
   val live = {
     implicit val trace = Trace.empty
-    val instance = new Live()
+    val instance       = new Live()
 
-      ZLayer.scoped(ZIO.acquireRelease(
+    ZLayer.scoped(
+      ZIO.acquireRelease(
         instance.reset("target/test-reports-zio/output.json") *>
           instance.makeOutputDirectory().orDie *>
           ZIO.succeed(instance)
       )(_ =>
         instance.removeTrailingComma("target/test-reports-zio/output.json") *>
           instance.closeJsonStructure("target/test-reports-zio/output.json") *>
-        ZIO.debug("Should remove last command and add closing square bracket")))
+          ZIO.debug("Should remove last command and add closing square bracket")
+      )
+    )
   }
 
   class Live() extends ExecutionEventPrinter {
     override def print(event: ExecutionEvent): ZIO[Any, Nothing, Unit] = {
       implicit val trace = Trace.empty
-//      event match {
-//        case ExecutionEvent.SectionEnd(labelsReversed, id, ancestors) =>
-//          removeTrailingComma("output.json")
-//
-//        case _ => ()
-//      }
       writeFile("target/test-reports-zio/output.json", jsonify(event)).orDie
     }
 
-    private def jsonify[E](test: Either[TestFailure[E], TestSuccess]): String = test match {
-      // TODO Render annotations
-      case Left(value) =>
-        "Failure"
-      case Right(value) =>
-        "Success"
-    }
+    private def jsonify[E](test: Either[TestFailure[E], TestSuccess]): String =
+      test match {
+        case Left(value) =>
+          "Failure"
+        case Right(value) =>
+          "Success"
+      }
 
-    private def jsonify(testAnnotationMap: TestAnnotationMap): String = {
-      TestAnnotationRenderer.default.run(List.empty, testAnnotationMap)
+    private def jsonify(testAnnotationMap: TestAnnotationMap): String =
+      TestAnnotationRenderer.default
+        .run(List.empty, testAnnotationMap)
         .map(s => s.replace("\"", "\\\""))
         .mkString(" : ")
-    }
 
     private def jsonify(executionEvent: ExecutionEvent): String = executionEvent match {
       case ExecutionEvent.Test(labelsReversed, test, annotations, ancestors, duration, id, fullyQualifiedName) =>
         s"""
-          |  {
-          |     "name" : "$fullyQualifiedName/${labelsReversed.reverse.mkString("/")}",
-          |     "status" : "${jsonify(test)}",
-          |     "durationMillis" : "${duration}",
-          |     "annotations" : "${jsonify(annotations)}"
-          |  },""".stripMargin
+           |  {
+           |     "name" : "$fullyQualifiedName/${labelsReversed.reverse.mkString("/")}",
+           |     "status" : "${jsonify(test)}",
+           |     "durationMillis" : "${duration}",
+           |     "annotations" : "${jsonify(annotations)}"
+           |  },""".stripMargin
       case ExecutionEvent.SectionStart(labelsReversed, id, ancestors) =>
         ""
-//        s"""{
-//           |   "suiteName" : "${labelsReversed.head}",
-//           |   "children" : [
-//           |""".stripMargin
       case ExecutionEvent.SectionEnd(labelsReversed, id, ancestors) =>
         // TODO Deal with trailing commas
         ""
-//        s"""
-//          |   ]
-//          |}${if(ancestors.head != SuiteId.global) "," else ""}""".stripMargin
       case ExecutionEvent.TopLevelFlush(id) => "TODO TopLevelFlush"
       case ExecutionEvent.RuntimeFailure(id, labelsReversed, failure, ancestors) =>
         "TODO RuntimeFailure"
     }
 
     def reset(path: => String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
-      ZIO.acquireReleaseWith(ZIO.attemptBlockingIO(new java.io.FileWriter(path, false)))(f =>
-        ZIO.attemptBlocking(f.close()).orDie
-      ) { f =>
-        ZIO.attemptBlockingIO(f.write("["))
-      }.orDie
+      ZIO
+        .acquireReleaseWith(ZIO.attemptBlockingIO(new java.io.FileWriter(path, false)))(f =>
+          ZIO.attemptBlocking(f.close()).orDie
+        ) { f =>
+          ZIO.attemptBlockingIO(f.write("["))
+        }
+        .orDie
 
     def closeJsonStructure(path: => String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
       writeFile(path, "]").orDie
 
     def makeOutputDirectory()(implicit trace: Trace) = ZIO.attempt {
-      import java.io.IOException
-      import java.nio.file.Files
-      import java.nio.file.Path
-      import java.nio.file.Paths
+      import java.nio.file.{Files, Paths}
 
       val fp = Paths.get("target/test-reports-zio/newfile.txt")
       Files.createDirectories(fp.getParent)
@@ -103,8 +94,7 @@ object ExecutionEventJsonPrinter {
 
     import java.io._
 
-
-    def removeTrailingComma(filePath: String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] = ZIO.succeed{
+    def removeTrailingComma(filePath: String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] = ZIO.succeed {
       try {
         val file = new RandomAccessFile(filePath, "rw")
         // Move the file pointer to the last character
@@ -114,7 +104,7 @@ object ExecutionEventJsonPrinter {
         do {
           c = file.read
           file.seek(file.getFilePointer - 2)
-        } while ( {
+        } while ({
           Character.isWhitespace(c)
         })
         // If the non-whitespace character is a comma, remove it
@@ -136,14 +126,14 @@ object ExecutionEventJsonPrinter {
     }
 
   def writeFile(path: => Path, content: => String)(implicit
-                                                   trace: Trace,
-                                                   d: DummyImplicit
+    trace: Trace,
+    d: DummyImplicit
   ): ZIO[Any, IOException, Unit] =
     writeFile(path.toString, content)
 
   def writeFileOutputStream(
-                             path: => String
-                           )(implicit trace: Trace): ZIO[Scope, IOException, ZOutputStream] =
+    path: => String
+  )(implicit trace: Trace): ZIO[Scope, IOException, ZOutputStream] =
     ZIO
       .acquireRelease(
         ZIO.attemptBlockingIO {
@@ -154,8 +144,8 @@ object ExecutionEventJsonPrinter {
       .map(_._2)
 
   def writeFileOutputStream(
-                             path: => Path
-                           )(implicit trace: Trace, d: DummyImplicit): ZIO[Scope, IOException, ZOutputStream] =
+    path: => Path
+  )(implicit trace: Trace, d: DummyImplicit): ZIO[Scope, IOException, ZOutputStream] =
     writeFileOutputStream(path.toString)
 
 }
